Add code from graph theory research
I wrote this code in Sage in May of 2019, and optimized it over the course of the research project so that we could efficiently iterate through large sets of graphs and find the ones we cared about. It was my first time using Python and Sage, as well as gt.sage (a library of graph theory objects and functions that my program relies on). If I were to write it again today, I'd look into Cythonizing it to speed it up further, although it served its purpose as-is well enough at the time.